refactor: CommandInputをコンポーネントに分離
